src/4_code/iot-kg-platform/app/infra/mqtt_client.py
import paho.mqtt.client as mqtt
from typing import Optional, Dict, Any
import json
import time


class IoTConnector:
    """
    IoT连接器接口实现（基于MQTT）
    Trace: [UC-01, UC-05, UC-07]
    """
    
    def __init__(self, broker: str, port: int, username: str, password: str):
        """
        初始化MQTT客户端
        Trace: [UC-01, UC-05, UC-07]
        
        Args:
            broker: MQTT broker地址
            port: MQTT端口
            username: MQTT用户名
            password: MQTT密码
        """
        self.broker = broker
        self.port = port
        self.username = username
        self.password = password
        
        # 创建MQTT客户端
        self.client = mqtt.Client()
        self.client.username_pw_set(username, password)
        
        # 连接MQTT broker
        self.client.connect(broker, port, 60)
        self.client.loop_start()
    
    def send_mqtt_message(self, topic: str, payload: Dict[str, Any]) -> bool:
        """
        发送MQTT消息
        Trace: [UC-05, UC-07]
        
        Args:
            topic: MQTT主题
            payload: 消息内容
        
        Returns:
            是否发送成功
        """
        try:
            message = json.dumps(payload)
            result = self.client.publish(topic, message)
            # 等待消息发送完成
            result.wait_for_publish()
            return result.rc == mqtt.MQTT_ERR_SUCCESS
        except Exception as e:
            print(f"发送MQTT消息失败: {e}")
            return False
    
    def revoke_certificate(self, cert_id: str) -> bool:
        """
        撤销设备证书
        Trace: [UC-07]
        
        Args:
            cert_id: 证书ID
        
        Returns:
            是否撤销成功
        """
        # 模拟证书撤销操作
        print(f"撤销证书: {cert_id}")
        # 实际项目中应调用证书管理服务
        return True
    
    def check_status(self, device_id: str) -> bool:
        """
        检查设备在线状态
        Trace: [UC-07]
        
        Args:
            device_id: 设备ID
        
        Returns:
            设备是否在线
        """
        # 模拟检查设备状态
        print(f"检查设备状态: {device_id}")
        # 实际项目中应通过MQTT或其他协议获取真实状态
        return True  # 模拟设备在线
    
    def send_command(self, device_id: str, command: str, params: Optional[Dict[str, Any]] = None) -> bool:
        """
        发送设备指令
        Trace: [UC-05, UC-07]
        
        Args:
            device_id: 设备ID
            command: 指令类型
            params: 指令参数
        
        Returns:
            是否发送成功
        """
        topic = f"devices/{device_id}/commands"
        payload = {
            "command": command,
            "timestamp": time.time()
        }
        if params:
            payload.update(params)
        
        return self.send_mqtt_message(topic, payload)
    
    def get_command_result(self, device_id: str, command_id: str) -> Dict[str, Any]:
        """
        获取指令执行结果
        Trace: [UC-07]
        
        Args:
            device_id: 设备ID
            command_id: 指令ID
        
        Returns:
            指令执行结果
        """
        # 模拟获取指令结果
        print(f"获取指令结果: {device_id}, {command_id}")
        # 实际项目中应实现真实的结果查询逻辑
        return {
            "status": "COMPLETED",
            "message": "指令执行成功"
        }
    
    def close(self):
        """
        关闭MQTT连接
        Trace: [UC-01, UC-05, UC-07]
        """
        self.client.loop_stop()
        self.client.disconnect()
