src/4_code/iot-kg-platform/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
neo4j==5.14.1
paho-mqtt==1.6.1
python-dotenv==1.0.0
